move general middleware into middlewares folder, add identity_verification.styl
